requirements.txt
PySide6>=6.7.0
PyYAML>=6.0.1
bacpypes3>=0.0.98; python_version < "3.13"
